Norme + modification fonctions
